tools/adhoc_mentorlink_autoclose/requirements.txt
gspread==6.1.4
google-auth==2.35.0
pandas==2.2.2
python-dateutil==2.9.0.post0
ruamel.yaml==0.18.15
Unidecode==1.3.8